Refactor registry pattern and organize graders into modular structure
Major refactoring to eliminate code duplication and improve organization:

## Registry Pattern Consolidation
- Create BaseRegistry class with common registry functionality
- Refactor GraderRegistry and AssignmentRegistry to inherit from BaseRegistry
- Eliminate ~80 lines of duplicate code between registries
- Add automatic module discovery and loading

## Modular Grader Organization
- Extract base grader classes (Grader, Grader__docker) to grader.py
- Create graders/ package with specialized implementations:
  * manual.py - Manual grading workflow
  * docker_graders.py - Configurable Docker graders
  * cst334.py - CST334-specific graders (regular and online)
  * step_by_step.py - Step-by-step comparison grader

## Benefits
- Reduced code duplication and improved maintainability
- Clear separation of concerns with each grader type in own file
- Better documentation and examples for different grader patterns
- Easier extensibility - new graders just need new files in graders/
- Maintains full backward compatibility
